+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 0.2.0-dev
 - only first line of sections TEST, ARGS and FILE_EXTERNAL is now parsed
+- sections GET and COOKIE are parsed now
 
 Version 0.1.0
 - initial version